add cleanup logs to integ test
