+++ agent/engine/execcmd/manager_unsupported.go
@@ -23,9 +23,9 @@ import (
 )
 
 const (
-	// HostLogDir here is used used while cleaning up exec logs when task exits.
+	// ECSAgentExecLogDir here is used used while cleaning up exec logs when task exits.
 	// When this path is empty, nothing is cleaned up for unsupported platforms.
-	HostLogDir = ""
+	ECSAgentExecLogDir = ""
 )
 
 // Note: exec cmd agent is a linux-only feature, thus implemented here as a no-op.
